Move test such that methods don't need public access.
